factor out backend.InventoryContext
